% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Plotting CountryPopulation.}
\usage{
CountryPopulation(x)
}
\arguments{
\item{x}{A real Country Name}
}
\value{
Population change of this country
}
\description{
Graphing the country population of what you enter
}
\details{
The goal is to show the population graph of the country you entered,
population changes from 1950-2020.
}
\examples{
CountryPopulation("Italy")
CountryPopulation("China")
}
