% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{World Cup Attendance Information.}
\format{
a data frame includes 22 World Cup brief information
\describe{
  \item{Matches}{Matches}
  \item{Totalattendance}{Total attendance}
  \item{Averageattendance}{Average attendance}
  \item{WorldCup}{Host Country and Hosting Year}
}
}
\source{
\url{ https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
World Cup Attendance Information.
}
\keyword{datasets}
