% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World Population from 1950 to 2020.}
\format{
a data frame includes 235 countries population data
\describe{
  \item{Countries}{Country Name.}
  \item{1950-2020}{Year}
}
}
\source{
\url{ http://creativecommons.org/licenses/by/3.0/igo/}
}
\usage{
WorldPopulation
}
\description{
World Population from 1950 to 2020.
}
\keyword{datasets}
